+++ magneto/src/lib.rs
@@ -1,4 +1,0 @@
-// This is the rust file containing the logic for the simulation.
-// 
-// Author: Brayden JoHantgen
-// Last Update: 4/27/2026